+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Casts/Tabs/OwnCastsByPartyMembers/Tabs/OwnCastsByPartyMember.lua
@@ -8,6 +8,7 @@ UI
 ]]
 
 local L = LibStub("AceLocale-3.0"):GetLocale("MyDungeonsBook");
+local AceGUI = LibStub("AceGUI-3.0");
 
 --[[--
 @param[type=Frame] parentFrame
@@ -22,6 +23,7 @@ function MyDungeonsBook:OwnCastsByPartyMemberFrame_Create(parentFrame, challenge
     local columns = self:OwnCastsByPartyMemberFrame_GetHeadersForTable(challengeId, unitId);
     local table = self:TableWidget_Create(columns, 10, 40, nil, ownCastsByPartyMemberFrame, "own-casts-by-" .. unitId);
     table:SetData(data);
+    table.frame:SetPoint("TOPLEFT", 260, -40);
     table:RegisterEvents({
         OnEnter = function (_, cellFrame, data, _, _, realrow, column)
             if (realrow) then
@@ -38,11 +40,11 @@ function MyDungeonsBook:OwnCastsByPartyMemberFrame_Create(parentFrame, challenge
             end
         end
     });
-    local summaryData = self:OwnCastsByPartyMemberFrame_GetDataForSummaryTable(challengeId, key, unitId);
+    local summaryData = self:OwnCastsByPartyMemberFrame_GetImportantDataForSummaryTable(challengeId, key, unitId);
     local summaryColumns = self:OwnCastsByPartyMemberFrame_GetHeadersForSummaryTable();
-    local summaryTable = self:TableWidget_Create(summaryColumns, 10, 40, nil, ownCastsByPartyMemberFrame, "own-casts-by-" .. unitId .. "-summary");
+    local summaryTable = self:TableWidget_Create(summaryColumns, 9, 40, nil, ownCastsByPartyMemberFrame, "own-casts-by-" .. unitId .. "-summary");
     summaryTable:SetData(summaryData);
-    summaryTable.frame:SetPoint("TOPLEFT", 380, -40);
+    summaryTable.frame:SetPoint("TOPLEFT", 0, -80);
     summaryTable:RegisterEvents({
         OnEnter = function (_, cellFrame, data, _, _, realrow, column)
             if (realrow) then
@@ -59,6 +61,17 @@ function MyDungeonsBook:OwnCastsByPartyMemberFrame_Create(parentFrame, challenge
             end
         end
     });
+    local allImportantCastsToggle = AceGUI:Create("CheckBox");
+    allImportantCastsToggle:SetLabel(L["Show All Casts"]);
+    allImportantCastsToggle:SetValue(false);
+    allImportantCastsToggle:SetCallback("OnValueChanged", function(_, _, newValue)
+        if (newValue) then
+            summaryTable:SetData(self:OwnCastsByPartyMemberFrame_GetAllDataForSummaryTable(challengeId, unitId));
+        else
+            summaryTable:SetData(self:OwnCastsByPartyMemberFrame_GetImportantDataForSummaryTable(challengeId, key, unitId));
+        end
+    end);
+    ownCastsByPartyMemberFrame:AddChild(allImportantCastsToggle);
     return ownCastsByPartyMemberFrame;
 end
 
@@ -189,14 +202,14 @@ function MyDungeonsBook:OwnCastsByPartyMemberFrame_GetDataForTable(challengeId, 
 end
 
 --[[--
-Map summary data about own casts by party member `unitId` for challenge with id `challengeId`.
+Map summary data about "important" own casts by party member `unitId` for challenge with id `challengeId`.
 
 @param[type=number] challengeId
 @param[type=string] key for mechanics table
 @param[type=unitId] unitId
 @return[type=table]
 ]]
-function MyDungeonsBook:OwnCastsByPartyMemberFrame_GetDataForSummaryTable(challengeId, key, unitId)
+function MyDungeonsBook:OwnCastsByPartyMemberFrame_GetImportantDataForSummaryTable(challengeId, key, unitId)
     local tableData = {};
     if (not challengeId) then
         return nil;
@@ -218,6 +231,43 @@ function MyDungeonsBook:OwnCastsByPartyMemberFrame_GetDataForSummaryTable(challe
         for _, _ in pairs(spellUsages) do
             usageCount = usageCount + 1;
         end;
+        tinsert(tableData, {
+            cols = {
+                {value = spellId},
+                {value = spellId},
+                {value = spellId},
+                {value = usageCount}
+            }
+        });
+    end
+    return tableData;
+end
+
+--[[--
+Map summary data about all own casts by party member `unitId` for challenge with id `challengeId`.
+
+@param[type=number] challengeId
+@param[type=unitId] unitId
+@return[type=table]
+]]
+function MyDungeonsBook:OwnCastsByPartyMemberFrame_GetAllDataForSummaryTable(challengeId, unitId)
+    local tableData = {};
+    if (not challengeId) then
+        return nil;
+    end
+    local challenge = self.db.char.challenges[challengeId];
+    local mechanics = challenge.mechanics["ALL-CASTS-DONE-BY-PARTY-MEMBERS"];
+    if (not mechanics) then
+        self:DebugPrint(string.format("No Own Casts Done By Party Members data for challenge #%s", challengeId));
+        return tableData;
+    end
+    local name, nameAndRealm = self:GetNameByPartyUnit(challengeId, unitId);
+    local mechanicsData = mechanics[name] or mechanics[nameAndRealm];
+    if (not mechanicsData) then
+        self:DebugPrint(string.format("No Own Casts found for %s or %s", name, nameAndRealm));
+        return tableData;
+    end
+    for spellId, usageCount in pairs(mechanicsData) do
         tinsert(tableData, {
             cols = {
                 {value = spellId},